requirements.txt
blinker==1.9.0
click==8.3.3
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paho-mqtt==2.1.0
Werkzeug==3.1.8
